+++ pages/three/cube.tsx
@@ -3,7 +3,7 @@ import { css } from "@emotion/core";
 import Link from "next/link";
 import DefaultLayout from "~/layouts/DefaultLayout";
 import { px } from "~/lib/cssUtil";
-import ThreeCube from "~/lib/ThreeCube";
+import SampleCubeAgent from "~/lib/SampleCubeAgent";
 import FullScreenThree from "~/components/FullScreenThree";
 
 const SWITCH_LINK = "/three/plane/";
@@ -17,10 +17,10 @@ const uiStyle = css({
 });
 
 export default () => {
-  const threeFullScreen = useMemo(() => new ThreeCube(), []);
+  const sampleCubeAgent = useMemo(() => new SampleCubeAgent(), []);
   return (
     <DefaultLayout>
-      <FullScreenThree agent={threeFullScreen} />
+      <FullScreenThree agent={sampleCubeAgent} />
       <div css={uiStyle}>
         <Link href={SWITCH_LINK}>
           <a href={SWITCH_LINK}>plane</a>
