+++ src/glsl/cubeGradFragment.glsl
